Added file safety with the EOSSDK backup

+++ BTD6EpicGamesModCompat/EOSSDK.cs
@@ -2,8 +2,7 @@
 
 namespace BTD6EpicGamesModCompat;
 
-internal static class EOSSDK
-{
+internal static class EOSSDK {
     // Paths for EOSSDK
     private const string EOSSDKName = "EOSSDK-Win64-Shipping.dll";
     private const string BTD6PluginsFolder = @"BloonsTD6_Data\Plugins\x86_64\";
@@ -12,19 +11,17 @@ internal static class EOSSDK
     private const string EOSSDKBackupPath = BackupFolder + EOSSDKName;
 
     // Remove EOSSDK to not crash melonloader immediately
-    public static void Remove()
-    {
+    public static void Remove() {
         Directory.CreateDirectory(BackupFolder);
         if (File.Exists(EOSSDKPath))
-            File.Move(EOSSDKPath, EOSSDKBackupPath);
+            File.Move(EOSSDKPath, EOSSDKBackupPath, true);
 
         Plugin.Logger.Msg("Removed EOSSDK");
     }
 
     // Restore EOSSDK for if the player wants to go unmodded
-    public static void Restore()
-    {
-        if (File.Exists(EOSSDKBackupPath))
+    public static void Restore() {
+        if (File.Exists(EOSSDKBackupPath) && !File.Exists(EOSSDKPath))
             File.Move(EOSSDKBackupPath, EOSSDKPath);
 
         Plugin.Logger.Msg("Restored EOSSDK");
